Update .gitignore + Fix base_type

+++ Common/include/TMIV/Common/Frame.hpp
@@ -39,35 +39,35 @@ namespace TMIV::Common {
 namespace detail {
 template <> struct PixelFormatHelper<YUV400P8> {
   static constexpr int nb_plane = 1;
-  using base_type = std::int16_t;
+  using base_type = std::uint8_t;
   static constexpr int getPlaneWidth(int, int W) { return W; }
   static constexpr int getPlaneHeight(int, int H) { return H; }
 };
 
 template <> struct PixelFormatHelper<YUV400P16> {
   static constexpr int nb_plane = 1;
-  using base_type = std::int16_t;
+  using base_type = std::uint16_t;
   static int getPlaneWidth(int, int W) { return W; }
   static int getPlaneHeight(int, int H) { return H; }
 };
 
 template <> struct PixelFormatHelper<YUV420P8> {
   static constexpr int nb_plane = 3;
-  using base_type = std::int16_t;
+  using base_type = std::uint8_t;
   static int getPlaneWidth(int id, int W) { return (id == 0) ? W : (W / 2); }
   static int getPlaneHeight(int id, int H) { return (id == 0) ? H : (H / 2); }
 };
 
 template <> struct PixelFormatHelper<YUV420P10> {
   static constexpr int nb_plane = 3;
-  using base_type = std::int16_t;
+  using base_type = std::uint16_t;
   static int getPlaneWidth(int id, int W) { return (id == 0) ? W : (W / 2); }
   static int getPlaneHeight(int id, int H) { return (id == 0) ? H : (H / 2); }
 };
 
 template <> struct PixelFormatHelper<YUV420P16> {
   static constexpr int nb_plane = 3;
-  using base_type = std::int16_t;
+  using base_type = std::uint16_t;
   static int getPlaneWidth(int id, int W) { return (id == 0) ? W : (W / 2); }
   static int getPlaneHeight(int id, int H) { return (id == 0) ? H : (H / 2); }
 };
